Python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
cuda-python
pypiwin32
numpy